meme_generator/requirements.txt
fastapi==0.115.0
python-multipart==0.0.12
Pillow==11.0.0
uvicorn==0.32.0
